small fixes and tests
