+++ landingzones/tests/test_permissions.py
@@ -17,9 +17,8 @@ from landingzones.tests.test_models import (
     ZONE_TITLE,
     ZONE_DESC,
 )
-from taskflowbackend.tests.base import (
-    TaskflowAPIPermissionTestBase,
-)
+from taskflowbackend.tests.base import TaskflowAPIPermissionTestBase
+
 
 # SODAR constants
 PROJECT_ROLE_OWNER = SODAR_CONSTANTS['PROJECT_ROLE_OWNER']
@@ -189,6 +188,182 @@ class TestLandingZonePermissions(TestLandingZonePermissionsBase):
         """Test ZoneCreateView with disabled non-superuser access"""
         url = reverse(
             'landingzones:create', kwargs={'project': self.project.sodar_uuid}
+        )
+        good_users = [self.superuser]
+        bad_users = [
+            self.user_owner_cat,
+            self.user_delegate_cat,
+            self.user_contributor_cat,
+            self.user_guest_cat,
+            self.user_finder_cat,
+            self.user_owner,
+            self.user_delegate,
+            self.user_contributor,
+            self.user_guest,
+            self.user_no_roles,
+            self.anonymous,
+        ]
+        self.assert_response(url, good_users, 200)
+        self.assert_response(url, bad_users, 302)
+
+    def test_zone_update(self):
+        """Test ZoneUpdateView permissions"""
+        self.landing_zone.status = 'ACTIVE'
+        self.landing_zone.save()
+        url = reverse(
+            'landingzones:update',
+            kwargs={'landingzone': self.landing_zone.sodar_uuid},
+        )
+        good_users = [
+            self.superuser,
+            self.user_owner_cat,
+            self.user_delegate_cat,
+            self.user_owner,
+            self.user_delegate,
+            self.user_contributor,
+        ]
+        bad_users = [
+            self.user_contributor_cat,
+            self.user_guest_cat,
+            self.user_finder_cat,
+            self.user_guest,
+            self.user_no_roles,
+            self.anonymous,
+        ]
+        self.assert_response(url, good_users, 200)
+        redirect_url = reverse(
+            'landingzones:list',
+            kwargs={'project': self.landing_zone.project.sodar_uuid},
+        )
+        self.assert_response(url, bad_users, 302, redirect_user=redirect_url)
+
+    def test_zone_update_archive(self):
+        """Test ZoneUpdateView with archived project"""
+        self.project.set_archive()
+        self.landing_zone.status = 'ACTIVE'
+        self.landing_zone.save()
+        url = reverse(
+            'landingzones:update',
+            kwargs={'landingzone': self.landing_zone.sodar_uuid},
+        )
+        good_users = [
+            self.superuser,
+        ]
+        bad_users = [
+            self.user_owner_cat,
+            self.user_delegate_cat,
+            self.user_contributor_cat,
+            self.user_guest_cat,
+            self.user_finder_cat,
+            self.user_owner,
+            self.user_delegate,
+            self.user_contributor,
+            self.user_guest,
+            self.user_no_roles,
+            self.anonymous,
+        ]
+        self.assert_response(url, good_users, 200)
+        redirect_url = reverse(
+            'landingzones:list',
+            kwargs={'project': self.landing_zone.project.sodar_uuid},
+        )
+        self.assert_response(url, bad_users, 302, redirect_user=redirect_url)
+
+    @override_settings(LANDINGZONES_DISABLE_FOR_USERS=True)
+    def test_zone_update_disable(self):
+        """Test ZoneUpdateView with disabled non-superuser access"""
+        self.landing_zone.status = 'ACTIVE'
+        self.landing_zone.save()
+        url = reverse(
+            'landingzones:update',
+            kwargs={'landingzone': self.landing_zone.sodar_uuid},
+        )
+        good_users = [
+            self.superuser,
+        ]
+        bad_users = [
+            self.user_owner_cat,
+            self.user_delegate_cat,
+            self.user_contributor_cat,
+            self.user_guest_cat,
+            self.user_finder_cat,
+            self.user_owner,
+            self.user_delegate,
+            self.user_contributor,
+            self.user_guest,
+            self.user_no_roles,
+            self.anonymous,
+        ]
+        self.assert_response(url, good_users, 200)
+        redirect_url = reverse(
+            'landingzones:list',
+            kwargs={'project': self.landing_zone.project.sodar_uuid},
+        )
+        self.assert_response(url, bad_users, 302, redirect_user=redirect_url)
+
+    def test_zone_delete(self):
+        """Test ZoneDeleteView permissions"""
+        self.landing_zone.status = 'ACTIVE'
+        self.landing_zone.save()
+        url = reverse(
+            'landingzones:delete',
+            kwargs={'landingzone': self.landing_zone.sodar_uuid},
+        )
+        good_users = [
+            self.superuser,
+            self.user_owner_cat,
+            self.user_delegate_cat,
+            self.user_owner,
+            self.user_delegate,
+            self.user_contributor,
+        ]
+        bad_users = [
+            self.user_contributor_cat,
+            self.user_guest_cat,
+            self.user_finder_cat,
+            self.user_guest,
+            self.user_no_roles,
+            self.anonymous,
+        ]
+        self.assert_response(url, good_users, 200)
+        self.assert_response(url, bad_users, 302)
+
+    def test_zone_delete_archive(self):
+        """Test ZoneDeleteView with archived project"""
+        self.landing_zone.status = 'ACTIVE'
+        self.landing_zone.save()
+        self.project.set_archive()
+        url = reverse(
+            'landingzones:delete',
+            kwargs={'landingzone': self.landing_zone.sodar_uuid},
+        )
+        good_users = [
+            self.superuser,
+            self.user_owner_cat,
+            self.user_delegate_cat,
+            self.user_owner,
+            self.user_delegate,
+            self.user_contributor,
+        ]
+        bad_users = [
+            self.user_contributor_cat,
+            self.user_guest_cat,
+            self.user_finder_cat,
+            self.user_guest,
+            self.user_no_roles,
+            self.anonymous,
+        ]
+        self.assert_response(url, good_users, 200)
+        self.assert_response(url, bad_users, 302)
+
+    @override_settings(LANDINGZONES_DISABLE_FOR_USERS=True)
+    def test_zone_delete_disable(self):
+        """Test ZoneDeleteView with disabled non-superuser access"""
+        self.landing_zone.status = 'ACTIVE'
+        self.landing_zone.save()
+        url = reverse(
+            'landingzones:delete',
+            kwargs={'landingzone': self.landing_zone.sodar_uuid},
         )
         good_users = [self.superuser]
         bad_users = [
